+++ src/components/accordion/CaseAccordion.tsx
