+++ src/components/card/CompanyCard.tsx
@@ -12,17 +12,13 @@ export interface CardProps {
 }
 
 const ESGTag: FC<{ grade: string | undefined }> = ({ grade }) => {
-  let display = grade
-  if (!grade || !Object.values(envGradeEnum).includes(grade)) {
-    display = 'N/A'
-  }
   return (
     <div
       className={`flex ${envGradeToColor(
         grade,
       )} header-2 aspect-square max-w-fit items-center rounded-xl text-center leading-8 text-white`}
     >
-      <div className="w-[109px] align-middle">{display}</div>
+      <div className="w-[109px] align-middle">{grade ? grade : 'N/A'}</div>
     </div>
   )
 }
