+++ src/components/chart/EmissionBarChart.tsx
