+++ src/components/display/Carousel/CarouselChild.tsx
