+++ src/components/index.ts
@@ -1,63 +1,63 @@
 // Buttons
-import AuthButton from './Buttons/AuthButton'
-import NavButton from './Nav/PrimaryNavBar/NavButton'
-import PrimaryButton from './Buttons/PrimaryButton'
-import ToTopButton from './Buttons/ToTopButton'
-import ReadMoreButton from './Buttons/ReadMoreButton'
-import BackButton from './Buttons/BackButton'
-import LoadMoreButton from './Buttons/LoadMoreButton'
-import TabButton from './Buttons/TabButton'
+import AuthButton from './button/AuthButton'
+import NavButton from './navigation/PrimaryNavBar/NavButton'
+import PrimaryButton from './button/PrimaryButton'
+import ToTopButton from './button/ToTopButton'
+import ReadMoreButton from './button/ReadMoreButton'
+import BackButton from './button/BackButton'
+import LoadMoreButton from './button/LoadMoreButton'
+import TabButton from './button/TabButton'
 // Card
-import CompanyCard from './Cards/CompanyCard'
-import AdminSelectCard from './Cards/AdminSelectCard'
-import DataCard from './Cards/DataCard'
-import LinkBox from './Cards/LinkBox'
+import CompanyCard from './card/CompanyCard'
+import AdminSelectCard from './card/AdminSelectCard'
+import DataCard from './card/DataCard'
+import LinkBox from './card/LinkBox'
 // Charts
-import EnergyRadialChart from './Charts/EnergyRadialChart'
-import LandingDonutChart from './Charts/LandingDonutChart'
-import FuelRadialChart from './Charts/FuelRadialChart'
-import EmissionBarChart from './Charts/EmissionBarChart'
-import FinanceBrushChart from './Charts/FinanceBrushChart'
+import EnergyRadialChart from './chart/EnergyRadialChart'
+import LandingDonutChart from './chart/LandingDonutChart'
+import FuelRadialChart from './chart/FuelRadialChart'
+import EmissionBarChart from './chart/EmissionBarChart'
+import FinanceBrushChart from './chart/FinanceBrushChart'
 // Displays
-import ListItem from './Displays/ListItem'
-import Carousel from './Displays/Carousel/Carousel'
-import Tabs from './Displays/Tabs/Tabs'
-import ClimateClock from './Displays/ClimateClock/ClimateClock'
-import ImageWithCaption from './Displays/Image/ImageWithCaption'
+import ListItem from './display/ListItem'
+import Carousel from './display/Carousel/Carousel'
+import Tabs from './display/Tabs/Tabs'
+import ClimateClock from './display/ClimateClock/ClimateClock'
+import ImageWithCaption from './display/Image/ImageWithCaption'
 // Accordions
-import RequestAccordion from './Accordions/RequestAccordion'
-import CaseAccordion from './Accordions/CaseAccordion'
-import ReadMoreAccordion from './Accordions/ReadMoreAccordion'
+import RequestAccordion from './accordion/RequestAccordion'
+import CaseAccordion from './accordion/CaseAccordion'
+import ReadMoreAccordion from './accordion/ReadMoreAccordion'
 // Nav
-import PrimaryNavBar from './Nav/PrimaryNavBar/PrimaryNavBar'
-import SecondaryNavBar from './Nav/SecondaryNavBar'
-import AdminNavBar from './Nav/AdminNavBar'
+import PrimaryNavBar from './navigation/PrimaryNavBar/PrimaryNavBar'
+import SecondaryNavBar from './navigation/SecondaryNavBar'
+import AdminNavBar from './navigation/AdminNavBar'
 // Selects
-import Select from './Selects/Select'
-import SearchBar from './Selects/SearchBar'
-import InviteAdminBar from './Selects/InviteAdminBar'
+import Select from './select/Select'
+import SearchBar from './select/SearchBar'
+import InviteAdminBar from './select/InviteAdminBar'
 // Text
-import NumberCircle from './Text/NumberCircle'
-import HighlightedTitle from './Text/HighlightedTitle'
-import CommentModal from './Modals/CommentModal'
-import SubmitRequestModal from './Modals/SubmitRequestModal'
-import CapitalizedTitle from './Text/CapitalizedTitle'
-import ShadowTitle from './Text/ShadowTitle'
-import StatusTag from './Text/StatusTag'
-import Tag from './Text/Tag'
+import NumberCircle from './text/NumberCircle'
+import HighlightedTitle from './text/HighlightedTitle'
+import CommentModal from './modal/CommentModal'
+import SubmitRequestModal from './modal/SubmitRequestModal'
+import CapitalizedTitle from './text/CapitalizedTitle'
+import ShadowTitle from './text/ShadowTitle'
+import StatusTag from './text/StatusTag'
+import Tag from './text/Tag'
 // Tables
-import AdminListTable from './Table/AdminListTable'
-import InvestmentTable from './Table/InvestmentTable'
-import RequestReviewTable from './Table/ReviewRequestTable'
+import AdminListTable from './table/AdminListTable'
+import InvestmentTable from './table/InvestmentTable'
+import RequestReviewTable from './table/ReviewRequestTable'
 //Tooltips
-import ToolTip from './Tooltips/ToolTip'
+import ToolTip from './tooltip/ToolTip'
 // Modals
-import GeneralModal from './Modals/GeneralModal'
-import LeaveConfirmationModal from './Modals/LeaveConfirmationModal'
+import GeneralModal from './modal/GeneralModal'
+import LeaveConfirmationModal from './modal/LeaveConfirmationModal'
 // Toast
-import Toast from './Toast/Toast'
+import Toast from './toast/Toast'
 // Footer
-import Footer from './Footer/Footer'
+import Footer from './footer/Footer'
 export {
   AuthButton,
   ListItem,
