+++ src/components/modal/GeneralModal.tsx
